Refactored: Move Tracing config settings to Tracing namespace
